rename: LyricsContentType에서 LyricsMimeType으로 변경
